requirements.txt
selene >=2.0.0b16
voluptuous == 0.13.1
pytest_voluptuous == 1.2.0
curlify ==2.2.1
Appium-Python-Client >= 2.7.1
pytest >= 7.2.0
pytest-xdist == 2.5.0
allure-pytest >= 2.12.0
requests == 2.28.1
pydantic == 1.10.2
python-dotenv == 0.21.0
selenium ~= 4.4.3
config == 0.5.1
allure-python-commons == 2.12.0
webdriver-manager == 3.8.5